hal/armv7m/imxrt: apply new structure of hal interface
JIRA: PD-210
