+++ hal/armv7m/hal.c
@@ -22,6 +22,9 @@ struct {
 hal_syspage_t *syspage;
 
 
+extern void _hal_cpuInit(void);
+
+
 void *hal_syspageRelocate(void *data)
 {
 	return data;
@@ -49,7 +52,7 @@ void _hal_start(void)
 void _hal_init(void)
 {
 	_hal_spinlockInit();
-//	_hal_exceptionsInit();
+	_hal_exceptionsInit();
 	_hal_interruptsInit();
 	_hal_cpuInit();
 	_hal_consoleInit();
