+++ hal/armv7m/imxrt/117x/console.c
@@ -17,10 +17,10 @@
 #define UART_CONSOLE 11
 #endif
 
-#include "cpu.h"
 #include "imxrt117x.h"
+#include "../../../cpu.h"
 #include "../../../console.h"
-#include "../../include/arch/imxrt1170.h"
+#include "../../../../include/arch/imxrt1170.h"
 
 #define CONCAT3(a, b, c) a##b##c
 #define CONSOLE_BAUD(n)  (CONCAT3(UART, n, _BAUDRATE))
@@ -107,9 +107,9 @@ void _hal_consoleInit(void)
 
 	/* Reset all internal logic and registers, except the Global Register */
 	*(console_common.uart + uart_global) |= 1 << 1;
-	hal_cpuDataBarrier();
+	hal_cpuDataMemoryBarrier();
 	*(console_common.uart + uart_global) &= ~(1 << 1);
-	hal_cpuDataBarrier();
+	hal_cpuDataMemoryBarrier();
 
 	/* Set baud rate */
 	t = *(console_common.uart + uart_baud) & ~((0x1f << 24) | (1 << 17) | 0xfff);
