+++ hal/armv7m/imxrt/_init.S
@@ -15,7 +15,7 @@
 
 #define __ASSEMBLY__
 
-#include "cpu.h"
+#include <arch/cpu.h>
 
 .syntax unified
 .cpu cortex-m7
