+++ hal/armv7m/stm32/interrupts.c
@@ -13,15 +13,13 @@
  * %LICENSE%
  */
 
-#include "interrupts.h"
-#include "spinlock.h"
-#include "cpu.h"
+#include "../../interrupts.h"
+#include "../../spinlock.h"
+#include "../../cpu.h"
 #include "pmap.h"
 #include "stm32.h"
 
 #include "../../../proc/userintr.h"
-
-#include "../../../include/errno.h"
 
 
 #if defined(CPU_STM32L152XD) || defined(CPU_STM32L152XE)
@@ -113,7 +111,7 @@ int hal_interruptsSetHandler(intr_handler_t *h)
 	spinlock_ctx_t sc;
 
 	if (h == NULL || h->f == NULL || h->n >= SIZE_INTERRUPTS)
-		return -EINVAL;
+		return -1;
 
 	hal_spinlockSet(&interrupts.spinlock, &sc);
 	h->got = hal_cpuGetGot();
@@ -126,7 +124,7 @@ int hal_interruptsSetHandler(intr_handler_t *h)
 	}
 	hal_spinlockClear(&interrupts.spinlock, &sc);
 
-	return EOK;
+	return 0;
 }
 
 
@@ -135,7 +133,7 @@ int hal_interruptsDeleteHandler(intr_handler_t *h)
 	spinlock_ctx_t sc;
 
 	if (h == NULL || h->f == NULL || h->n >= SIZE_INTERRUPTS)
-		return -EINVAL;
+		return -1;
 
 	hal_spinlockSet(&interrupts.spinlock, &sc);
 	_intr_remove(&interrupts.handlers[h->n], h);
@@ -145,7 +143,7 @@ int hal_interruptsDeleteHandler(intr_handler_t *h)
 
 	hal_spinlockClear(&interrupts.spinlock, &sc);
 
-	return EOK;
+	return 0;
 }
